tcs_loop

Call t_coffee to calculate Transitive Consistency Scores (TCS) for all alignments in a folder

tcs_loop(folder, pattern = NULL, number_cores = 1, other_args = NULL,
  ...)

Arguments

folder
Character vector of length one; the path to the folder containing the alignments.

pattern
An optional regular expression. Only alignment files with names that match the regular expression will be included.

number_cores
Number of cores for t-coffee to use.

other_args
Character vector; other arguments to pass to t_coffee. Each should be an element of the vector.

...
Additional other arguments. Not used by this function, but meant to be used by drake_plan for tracking during workflows.

Value

A numeric vector; the overall TCS scores for all alignments in the folder.

Externally, plain text files that include TCS scores for each sequence, position, and the whole alignment overall, for each alignment. The files will be written to folder, and named <alignment>.score_ascii, where <alignment> is the filename of the input alignment.

References

Chang JM, Di Tommaso P, Notredame C. TCS: a new multiple sequence alignment reliability measure to estimate alignment accuracy and improve phylogenetic tree reconstruction. Mol Biol Evol. 2014 Jun. 31(6):1625-37. http://www.tcoffee.org/Projects/tcs/

Examples

# NOT RUN {
tcs_loop(folder = "some/folder/with/alignments/", pattern = "fa.mafft.aln-cln$")
# }