Find long open reading frames.

Extracts long open reading frames (ORFs) from a fasta file containing transcript sequences (i.e., the transcriptome). This is a wrapper that calls TransDecoder.LongOrfs.

transdecoder_long_orfs(transcriptome_file, wd = here::here(),
  other_args = NULL, echo = pkgconfig::get_config("baitfindR::echo",
  fallback = FALSE), ...)

Arguments

transcriptome_file
Character vector of length one; the path to the fasta file containing transcript sequences (i.e., the transcriptome).

wd
Character vector of length one; the directory where the command will be run, and the output folder created.

other_args
Character vector; other arguments to pass to TransDecoder. Each should be an element of the vector.

echo
Logical; should the standard output and error be printed to the screen?

...
Additional other arguments. Not used by this function, but meant to be used by drake_plan for tracking during workflows.

Value

Within the R environment, a list with components specified in run.

Externally, an output folder in the working directory containing base frequencies and .cds, .gff3, and .pep files for the recovered ORFs. The folder will be named <transcriptome>.transdecoder_dir, where <transcriptome> is the value of that argument.

References

http://transdecoder.github.io

Examples

# NOT RUN {
library(ape)
temp_dir <- tempdir()
data("example_transcriptomes")
PSKY <- example_transcriptomes$PSKY
write.FASTA(PSKY, fs::path(temp_dir, "PSKY.fasta"))
transdecoder_long_orfs(
  transcriptome_file = fs::path(temp_dir, "PSKY.fasta"),
  wd = temp_dir
  )
list.files(temp_dir)
# }