build_blast_db

Wrapper to call makeblastdb

build_blast_db(in_seqs, out_name, db_type = "nucl", other_args = NULL,
  ...)

Arguments

in_seqs
Character vector of length one; the path to the fasta file containing the sequences to be included in the database.

out_name
Character vector of length one; name of BLAST database to be created.

db_type
Character vector of length one; "nucl" for DNA or "prot" for amino acids (proteins).

other_args
Character vector; other arguments to pass on to makeblastdb. For a list of options, run makeblastdb -help.

...
Additional other arguments. Not used by this function, but meant to be used by drake_plan for tracking during workflows.

Value

A series of files starting with out_name and ending in .phr, .pin, .pog, .psd, .psi, psq (for proteins) or .nhr, .nin, .nog, .nsd, .nsi, and .nsq (for DNA) that constitute the BLAST database in the directory specified by wd.

Details

Takes a list of input sequences and makes them into a BLAST database. Assumes that all BLAST commands are in the user's $PATH.

References

https://www.ncbi.nlm.nih.gov/books/NBK279690/