Run mcl.

This is a wrapper for the Markov Cluster Algorithm (mcl), a clustering algorithm for graphs. Here, it is meant to be used on genetic distances from BLAST.

mcl(mcl_input, mcl_output, i_value, e_value, other_args = NULL, ...)

Arguments

mcl_input
Character vector of length one; the path to the input file for mcl clustering.

mcl_output
Character vector of length one; the path to the output file produced by the mcl algorithm.

i_value
Numeric or character vector of length one; the inflation value.

e_value
Numeric or character vector of length one; the minimal -log transformed evalue to be considered by the algorithm.

other_args
Character vector; other arguments to pass to mcl. Each should be an element of the vector. For example, to pass "-abc" to specify the input file format and "--te" to specify number of threads, use c("--abc", "-te", "2").

...
Other arguments. Not used by this function, but meant to be used by drake_plan for tracking during workflows.

Value

A plain text file of tab-separated values, where each value on a line belongs to the same cluster.

References

Stijn van Dongen, A cluster algorithm for graphs. Technical Report INS-R0010, National Research Institute for Mathematics and Computer Science in the Netherlands, Amsterdam, May 2000. https://micans.org/mcl/

Examples

# NOT RUN {
mcl(mcl_input = "some/folder/distance.file", mcl_output = "some/folder/mcl_output.txt", i_value = 1.4, evalue = 5)
# }