Predict coding regions.

This is a wrapper for transdecoder.predict. It must be run in the same directory containing the .transdecoder_dir output folders from transdecoder_long_orfs. Optionally include the results of a blastp search to make sure that peptides with a blastp hit against the reference database are retained in the TransDecoder output.

transdecoder_predict(transcriptome_file, blast_result = NULL,
  wd = here::here(), other_args = NULL,
  echo = pkgconfig::get_config("baitfindR::echo", fallback = FALSE), ...)

Arguments

transcriptome_file
Character vector of length one; the path to the fasta file containing transcript sequences (i.e., the transcriptome).

blast_result
Character vector of length one; the path to the tab-separated text file containing the results from a blastp search of the transcriptome against a reference blast protein database. For the blast search, the output format should specified as: -outfmt 6.

wd
Character vector of length one; the directory where the command will be run. Must contain .transdecoder_dir folder with results from transdecoder_long_orfs.

other_args
Character vector; other arguments to pass to TransDecoder. Each should be an element of the vector.

echo
Logical; should the standard output and error be printed to the screen?

...
Additional other arguments. Not used by this function, but meant to be used by drake_plan for tracking during workflows.

Value

Within the R environment, a list with components specified in run.

Externally, four output files (.cds, .pep, .bed, and .gff3) for each input transcriptome file will be written to the working directory.

Details

For a more detailed example, see vignettes.

References

http://transdecoder.github.io

Examples

# NOT RUN {
transdecoder_predict("some/transcriptome_file.fa", "some/blast_result.txt")
# }